move workflow ro-crate generator to rocrate module
